Add enemy info popup
