+++ src/arena/views/infobar.rs
@@ -8,7 +8,7 @@ use super::view_components::{Frame, FrameKind};
 use super::View;
 use crate::after_image::{FontColor, FontSize, IconLoader, RenderCanvas, RenderContext, TextRenderer};
 use crate::atlas::{BoxResult, EasyECS};
-use crate::clash::{find_enemies, find_player, AmmoKind, CharacterInfoComponent, SkillResourceComponent, StatusComponent};
+use crate::clash::{find_enemies, find_player, summarize_character, AmmoKind, CharacterInfoComponent, SkillResourceComponent, StatusComponent};
 
 pub struct InfoBarView {
     position: SDLPoint,
@@ -44,55 +44,7 @@ impl InfoBarView {
     }
 
     fn render_character(&self, canvas: &mut RenderCanvas, ecs: &World, entity: Entity, offset: &mut i32, show_status_effect: bool) -> BoxResult<()> {
-        let char_infos = &ecs.read_storage::<CharacterInfoComponent>();
-        let char_info = char_infos.grab(entity);
-        let defenses = &char_info.character.defenses;
-        let health_text = {
-            if defenses.absorb != 0 {
-                format!("Health: (+{:.2}) {:.2}/{:.2}", defenses.absorb, defenses.health, defenses.max_health)
-            } else {
-                format!("Health: {:.2}/{:.2}", defenses.health, defenses.max_health)
-            }
-        };
-        self.small_text(canvas, health_text.as_str(), offset)?;
-
-        if defenses.max_dodge != 0 {
-            self.small_text(canvas, format!("Dodge : {:.2}/{:.2}", defenses.dodge, defenses.max_dodge).as_str(), offset)?;
-        }
-        if defenses.armor != 0 {
-            self.small_text(canvas, format!("Armor: {:.2}", defenses.armor).as_str(), offset)?;
-        }
-
-        let resources = &ecs.read_storage::<SkillResourceComponent>();
-        if let Some(resource) = resources.get(entity) {
-            self.small_text(canvas, format!("Exhaustion: {:.2}", resource.exhaustion).as_str(), offset)?;
-
-            self.small_text(canvas, format!("Focus: {:.2}", resource.focus).as_str(), offset)?;
-
-            for kind in AmmoKind::into_enum_iter() {
-                match resource.max.get(&kind) {
-                    Some(value) => {
-                        self.small_text(canvas, format!("{:?}: {:.2}/{:.2}", kind, resource.ammo[&kind], value).as_str(), offset)?;
-                    }
-                    None => {}
-                }
-            }
-        }
-
-        let temperature = char_info.character.temperature.current_temperature();
-        if temperature != 0 {
-            self.small_text(canvas, format!("Temperature: {:.2}", temperature).as_str(), offset)?;
-        }
-
-        if show_status_effect {
-            let statuses = &ecs.read_storage::<StatusComponent>();
-            if let Some(status) = statuses.get(entity) {
-                let all = status.status.get_all_display_status();
-                if !all.is_empty() {
-                    self.small_text(canvas, &format!("Status: {}", all.join(" ")), offset)?;
-                }
-            }
-        }
+        summarize_character(ecs, entity, show_status_effect, |t| self.small_text(canvas, t, offset).unwrap());
         Ok(())
     }
